After encountering the latest Alpha snapshot and discovering @RPicster's great efforts on the Particle Turbulence PR, I found some areas for improvement and rework, to hopefully start this feature off on the best footing possible for the upcoming 4.0 release. (Updated description here following discussions further down)

The current implementation merged in #55387 uses directionally-scrolling 3D noise to create an evolving pattern in 3D. This is problematic because it forces the consideration of a scrolling direction, which creates unavoidable and often unfitting bias in resulting patterns. The included speed-random parameter is poised as a means to conceal these effects, but it's only a moderately-effective patch on top of a problem that doesn't need to exist in the first place. This PR replaces the scrolling 3D noise with 4D noise which varies in place over time.

Further, the included formula for curl noise is highly complex and is based on axis-biased arithmetic of uncertain mathematical origin. This PR replaces it with a more concise solution that uses fewer noise calls while also producing more natural effects.

_These changes do break turbulence configs made using the existing dev snapshots!_

### Full change set:

- Exchanged scrolling 3D noise pattern variation approach with 4D noise that evolves in place, producing more convincing effects through more intuitive configuration. Replaced `turbulence_noise_speed` (vec3) with `turbulence_noise_evolution_speed` (float). ~~2D particles use evolving 3D noise.~~
  - I can also re-add `turbulence_noise_speed` (vec3) if it is decided that the scrolling is useful -- perhaps as `turbulence_noise_directional_trend` to be more explicit about its purpose and effect.
- Replaced the then-merged 3D curl formula with a simple and orientation-invariant `cross(noiseGradientA, noiseGradientB)` which produces a proper divergence-free flow field with many fewer steps.
-~~Added a separate 2D case which uses an even simpler `rotate90(noiseGradient)` formula (paraphrased code).~~
- Replaced displaced-sampling-based noise gradient (partial derivative vector) computation with single-evaluation analytic derivative sampling.
- Replaced Perlin noise algorithm with more-directionally-unbiased-looking Simplex noise. Simplex is also easier to calculate the analytic derivatives for.
  - I used two octaves of noise per each gradient computation to remove / vary up the straight line segments some of the particles were following in the single layers. It also increases variation.
- Adjusted the curve of `turbulence_noise_scale` as it's passed to the shader, to better reflect the description in the documentation. `shader_turbulence_noise_scale = 1.0 / turbulence_noise_scale;`
- Removed `turbulence_noise_speed_random` since it no longer has its necessary use case.
- ~~Removed `turbulence_noise_strength` since it had no effect on the particle update step, and its effect during the initialization step could be accomplished just the same by increasing `turbulence_initial_displacement_min` and `turbulence_initial_displacement_max`.~~ Determined that it does produce an effect and re-added corresponding functionality though a `turbulence_sharpness` parameter.
  - It was ~~also~~ previously used in the definition of `adj_contrast, which no longer plays a role now that the partial derivative vector is computed analytically.
- Changed existing formula for `turbulence_influence` (from values sampled according to `turbulence_influence_min` and `turbulence_influence_max`), which biases the result towards more influence, to do so more cleanly. [old](https://github.com/RPicster/godot/blob/9137ba79269cd957753fce8d783ffb01c8ac60fa/scene/resources/particles_material.cpp#L706), [new](https://github.com/KdotJPG/godot/blob/e33c4e1131bdb71ef5f5edcee8b73427e5ffdc94/scene/resources/particles_material.cpp#L848-L849) https://www.desmos.com/calculator/hmsaly3f4b
  - ~~I also adjusted the way the normalization happens so that the new vector is always the same length as the old.~~
- ~~Fixed the distance-from-center selection in the existing `EMISSION_SHAPE_SPHERE` code -- it needs to be passed through a cube-root for a proper uniform distribution that doesn't cluster at the center.~~
  - ~~This technically does change existing patterns, but I wouldn't generally expect it to break any. If I should revert it or implement it differently, just give me the word.~~

### ~~Motion to also remove Initial Displacement~~

~~Initial Displacement seems well-intentioned, but I haven't found any behavior in it that I couldn't replicate with the Emission Shape features. If there is consensus, I will update this PR with its removal. (@RPicster feel free to fill me in if I'm missing something.)~~

---

![image](https://user-images.githubusercontent.com/8829856/185583392-755d7cae-f32e-4b17-8c6b-2668f6389934.png)

---

All code added is either MIT-licensed or self-authored.