@RPicster 2022-08-19
I'm in vacation for two weeks and won't have the time to properly review this PR.

I saw that there are a lot of changes and if it's ok I would like to review it in practice before big changes like this are getting merged.

There are a couple of things I found on the first look where I have some "feelings". But I like to properly review it and test it before saying anything.
I think I would rather have easier to maintain code instead of optimizing too much - talking about separate noise functions for 2D and 3D here. Let's please keep it 3D focused. 
Especially because I have some things in the works to "fake" some 3D behaviour in 2D in the future and there, the Z translation would be necessary in 2D too.

One thing that I can tell already is, that the changes to the emission shapes should go into a separate PR.

And, please also keep in mind the necessary license changes in this PR. Where are the functions from? What is necessary for Godot and Godot users to respect the license?