@KdotJPG 2022-08-19
No worries. Thanks for taking the time to give the initial comment!

I went ahead and removed the 2D-specific code, as well as the changes to emission shape. I'll set up a second PR for them at some point.

All the code is either MIT or my own. To break it down:

- The base for Simplex noise 4D comes from the following file in its MIT-licensed repository: [noise4D.glsl](https://github.com/ashima/webgl-noise/blob/master/src/noise4D.glsl) and its output partial derivative/gradient vector code is adapted from this other file there: [noise3Dgrad.glsl](https://github.com/ashima/webgl-noise/blob/master/src/noise3Dgrad.glsl)
- The hash functions with replace the `mod289`/`permute` randomness come from [this Shadertoy here](https://www.shadertoy.com/view/4djSRW), which declares itself with the MIT-license ([webarchive link for ongoing verifiability](https://web.archive.org/save/https://www.shadertoy.com/view/4djSRW)). (white noise audio warning)

The new vertex gradient system between the hash and the noise code itself is self-authored. The 4D noise in the repository seemed to produce some diagonal patterns that changed when I rendered slices of the noise with different coordinate permutations, meaning the gradient directions were not well-distributed. The replacement here first distributes the points on the surface of a tesseract, then uses a two-stages of polynomial approximation to get something like the result of the 4D version of [this](https://catlikecoding.com/unity/tutorials/cube-sphere/#MathJax-Element-42-Frame). The 'magic numbers' come from integral regression solving in Desmos.

I also realized I could re-tune the `turbulence_noise_scale` curve to be more in line with the description, and perhaps what users might expect from that parameter. It's a simple division now, making the term 'scale' define a number proportional to the average distance between bumps and valleys in the noise.

Feel free to let me know of any more questions / concerns when they come to you.