@KdotJPG 2022-08-20
All good! Figured it was that the moment I saw the cut-off comment.

Re: noise code complexity, I see where you're coming from, and definitely agree there's a learning curve to understanding the internals of this type of noise. What I do see as helping its cause in that regard is that it's fairly self-contained. While it takes space as a block of code that needs study to completely understand, in general you don't need to fully grasp it to write good code around it. I wouldn't consider it worth settling for a quality downgrade for a more-immediately-intuitive function here -- or really in most cases when paired with decent encapsulation principles, but I would certainly agree the code's readability as it is can be improved.

- I've gone ahead and revised the noise code with clearer variable names and more verbose comments.