@Zireael07 2022-08-20
@clayjohn: Noise, in my experience, is the sort of thing you implement once and don't care afterwards. It's better that it's more performant even at the cost of readability. If you really want people to understand how noise is done, you could link in a comment to a more readable, unoptimized implementation "if you want to understand noise look it up at blah blah"