@RPicster 2022-08-20
> @clayjohn: Noise, in my experience, is the sort of thing you implement once and don't care afterwards. It's better that it's more performant even at the cost of readability. If you really want people to understand how noise is done, you could link in a comment to a more readable, unoptimized implementation "if you want to understand noise look it up at blah blah"

By the way, can you provide real life comparisons of particle systems with a performance gain that can give a good reason to change the simple functions with the more complex ones?

You pointed out that my original approach uses more calls, but how do the two approaches compare in terms of real world performance? It would be very interesting to see what differences we're talking about here.