@Zireael07 2022-08-20
Unsure if you're asking me or clayjohn, as I haven't used the turbulence (yet) but I plan on using it to make my tire smoke less uniform and performance will definitely be a concern.

From my experience testing/looking up shaders on Shadertoy, the number of noise lookups can **make or break** a shader. For instance, if you look up 'gas giant' or 'planet', many of those have very many lookups even to the point of freezing/crashing webgl. When implementing my own gas giant shader, I went for less 'realistic' but less lookups because I fully expect star systems with multiple giants and I don't want my performance to suffer (it's already suffering because in 3.5 there are no instance uniforms/parameters, so to have different individual planets I need to make them unique/local to scene so no instancing)