@RPicster 2022-08-20
I just want to avoid changing things when there is no problem ( https://docs.godotengine.org/en/stable/community/contributing/best_practices_for_engine_contributors.html ).

Me and several other people tested the turbulence with millions of particles and never had any performance problems - so it would be great to see the real benefit before changing things "just because".