@Calinou 2023-04-29
**Testing project:** [test_particle_turbulence.zip](https://github.com/godotengine/godot/files/11360224/test_particle_turbulence.zip)

I've tested this PR rebased against `master` in a fully optimized editor build (`optimize=speed lto=full`). Overall, this PR seems to have no noticeable performance impact – there are minor differences here and there, but they're within margin of error. Very high particle counts (1 million) seem to perform very slightly better, but this may also be due to testing variance.

## Benchmark

**OS:** Fedora 37
**CPU:** Intel Core i9-13900K
**GPU:** GeForce RTX 4090 (NVIDIA 530.41.03)

### 1152×648

Amount | Turbulence OFF | Turbulence *ON* (`master`) | Turbulence *ON* (this PR)
-|-|-|-
**10** | 7962 FPS (0.13 mspf) | 5210 FPS (0.19 mspf) | 5232 FPS (0.19 mspf)
**100** | 7855 FPS (0.13 mspf) | 5210 FPS (0.19 mspf) | 5208 FPS (0.19 mspf)
**1,000** | 7836 FPS (0.13 mspf) | 5197 FPS (0.20 mspf) | 5189 FPS (0.20 mspf)
**10,000** | 7475 FPS (0.13 mspf) | 4999 FPS (0.20 mspf) | 4994 FPS (0.20 mspf)
**100,000** | 4882 FPS (0.20 mspf) | 3649 FPS (0.27 mspf) | 3757 FPS (0.27 mspf)
**1,000,000** | 950 FPS (1.05 mspf) | 889 FPS (1.13 mspf) | 920 FPS (1.09 mspf)

### 3840×2160

Amount | Turbulence OFF | Turbulence *ON* (`master`) | Turbulence *ON* (this PR)
-|-|-|-
**10** | 5458 FPS (0.18 mspf) | 1397 FPS (0.72 mspf) | 1401 FPS (0.71 mspf)
**100** | 5440 FPS (0.18 mspf) | 1383 FPS (0.73 mspf) | 1392 FPS (0.72 mspf)
**1,000** | 5437 FPS (0.18 mspf) | 1380 FPS (0.73 mspf) | 1392 FPS (0.72 mspf)
**10,000** | 5254 FPS (0.19 mspf) | 1375 FPS (0.73 mspf) | 1379 FPS (0.73 mspf)
**100,000** | 3664 FPS (0.28 mspf) | 1239 FPS (0.80 mspf) | 1244 FPS (0.80 mspf)
**1,000,000** | 844 FPS (1.18 mspf) | 575 FPS (1.74 mspf) | 591 FPS (1.69 mspf)

It's interesting to see the mspf differential becomes much greater at higher resolutions. I thought particle shaders' performance weren't influenced by viewport resolution, since they run once for every particle and not for each pixel. Besides, in the testing project, each particle is only a few pixels tall at 3840×2160.